Move more code out of controller to sessions
